bugs with types were fixed

+++ src/Dijkstra.cpp
@@ -6,7 +6,7 @@
 #define pair std::pair
 #define vector std::vector
 vector<int> Dijkstra(vector < vector < pair<int, int> > > g, int start) {
-    if ( g.empty() || (start < 1) || (start > g.size()) ) throw 1;
+    if ( g.empty() || (start < 1) || (start > static_cast<int>(g.size())) ) throw 1;
     vector<int> d(g.size(), INF);
     int v;
     d[start-1] = 0;
@@ -16,7 +16,7 @@ vector<int> Dijkstra(vector < vector < pair<int, int> > > g, int start) {
         pair<int, int> tmp = *s.begin();
         s.erase(s.begin());
         v = tmp.second-1;
-        for (int i = 0; i < g[v].size(); ++i) {
+        for (int i = 0; i < static_cast<int>(g[v].size()); ++i) {
             if (d[v] + g[v][i].second < d[g[v][i].first-1]) {
                 if (d[g[v][i].first-1] != INF)
                     s.erase(s.find(pair<int, int>(d[g[v][i].first-1],
